auth_service/requirements.txt
alembic==1.4.2
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Faker==4.0.3
flasgger==0.9.4
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
mistune==0.8.4
# pkg-resources==0.0.0
PyJWT==1.4.2
PyMySQL==0.9.3
pyrsistent==0.16.0
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
